Use add_suggested_values_to_schema for showing form
Don't use the schema_builder function with the raw options data because optional fields may not exist meaning the schema creation function fails as there is no default.

Instead use the static schema with defaults as the basis and then add_suggested_values_to_schema to replace the defaults with the existing values.