Add Defaults for Missing Optional Config Entries
If a loaded config is missing an value that its schema marks as optional, then automatically add in the default. This helps in cases where new optional parameters are added but the stored configuration is from an old version.
This also removes the need for the ugly hack of manually setting the new `invert_logic` fields in Light and PWM to False if missing.