Brightness Calibration Factor for RGB LEDs
RGB LEDs can have different brightness for each colour meaning all on is not really white. To correct for this it would be useful to be able to apply a calibration factor for each colour to normalise the brightness across each of the colours.

This commit adds three new configuration parameters for the RGB LED device which are 0-100% values for maximum brightness of each colour. These are applied as a scale factor for each colours PWM setting.